mcp_server/requirements.txt
fastapi==0.111.0
uvicorn
httpx
supabase
google-generativeai
python-dotenv
pydantic==2.4.2
pydantic-settings==2.0.3
PyPDF2==3.0.1
python-multipart>=0.0.7
bcrypt==4.0.1
python-jose[cryptography]==3.3.0

# File processing and text extraction
python-magic  # Advanced MIME type detection
python-docx   # Word document processing
openpyxl      # Excel file processing
lxml          # XML parsing
beautifulsoup4  # HTML parsing

# Semantic embeddings (newer versions with pre-built wheels for Windows)
sentence-transformers==2.7.0  # Semantic embeddings and re-ranking
torch==2.2.0  # Required by sentence-transformers
transformers>=4.40.0
numpy>=1.24.0  # Required for embeddings

